implementacao de banco de dados nas queixas

+++ disque-saude-backend-master/src/main/java/com/ufcg/si1/controller/QueixaController.java
@@ -51,7 +51,7 @@ public class QueixaController {
 	}
 
 	@RequestMapping(value = "/queixa/{id}", method = RequestMethod.GET)
-	public ResponseEntity<?> consultarQueixa(@PathVariable("id") long id) {
+	public ResponseEntity<?> consultarQueixa(@PathVariable("id") Integer id) {
 
 		Queixa queixaEncontrada = queixaService.findById(id);
 		if (queixaEncontrada == null) {
@@ -61,7 +61,7 @@ public class QueixaController {
 	}
 
 	@RequestMapping(value = "/queixa/{id}", method = RequestMethod.PUT)
-	public ResponseEntity<?> atualizaQueixa(@PathVariable("id") long id, @RequestBody Queixa queixa) {
+	public ResponseEntity<?> atualizaQueixa(@PathVariable("id") Integer id, @RequestBody Queixa queixa) {
 		try {
 			queixaService.atualizaQueixa(id, queixa);
 		} catch (Exception e) {
@@ -74,7 +74,7 @@ public class QueixaController {
 	}
 
 	@RequestMapping(value = "/queixa/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<?> excluiQueixa(@PathVariable("id") long id) {
+	public ResponseEntity<?> excluiQueixa(@PathVariable("id") Integer id) {
 		try {
 			queixaService.excluiQueixaPorId(id);
 			} catch (Exception e) {
